requirements.txt
absl-py==2.1.0
accelerate==0.30.1
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
apex==0.1
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.0.9
certifi==2024.6.2
charset-normalizer==2.0.4
click==8.1.7
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
deepspeed==0.14.2
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.13.1
flash-attn==2.5.8
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
hjson==3.1.0
huggingface-hub==0.23.1
idna==3.7
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
kiwisolver==1.4.5
latexcodec==3.0.0
lmdb==1.4.1
MarkupSafe==2.1.3
matplotlib==3.9.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-collections==0.1.1
monty==2024.5.15
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
packaging==23.2
palettable==3.3.3
pandas==2.2.2
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
plotly==5.22.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybtex==0.24.0
pydantic==2.7.1
pydantic_core==2.18.2
pymatgen==2024.5.1
pynvml==11.5.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rdkit==2023.9.6
regex==2024.5.15
requests==2.32.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.3.0
setproctitle==1.3.3
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
spglib==2.4.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
torch_geometric==2.5.3
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.1
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
uncertainties==3.1.7
unicore==0.0.1
unimol_tools==1.0.0
urllib3==2.2.1
wandb==0.17.0
wheel==0.43.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
